+++ expenses/src/sdk/configuration.ts
@@ -49,7 +49,7 @@ export class Configuration {
         const contentType: string = httpRes?.headers?.["content-type"] ?? "";
 
         if (httpRes?.status == null) throw new Error(`status code not found in response: ${httpRes}`);
-        const res: operations.GetCompanyConfigurationResponse = {statusCode: httpRes.status, contentType: contentType};
+        const res: operations.GetCompanyConfigurationResponse = {statusCode: httpRes.status, contentType: contentType, rawResponse: httpRes};
         switch (true) {
           case httpRes?.status == 200:
             if (utils.matchContentType(contentType, `application/json`)) {
@@ -109,7 +109,7 @@ export class Configuration {
         const contentType: string = httpRes?.headers?.["content-type"] ?? "";
 
         if (httpRes?.status == null) throw new Error(`status code not found in response: ${httpRes}`);
-        const res: operations.SaveCompanyConfigurationResponse = {statusCode: httpRes.status, contentType: contentType};
+        const res: operations.SaveCompanyConfigurationResponse = {statusCode: httpRes.status, contentType: contentType, rawResponse: httpRes};
         switch (true) {
           case httpRes?.status == 200:
             if (utils.matchContentType(contentType, `application/json`)) {