+++ expenses/src/sdk/expenses.ts
@@ -62,7 +62,7 @@ export class Expenses {
         const contentType: string = httpRes?.headers?.["content-type"] ?? "";
 
         if (httpRes?.status == null) throw new Error(`status code not found in response: ${httpRes}`);
-        const res: operations.CreateExpenseDatasetResponse = {statusCode: httpRes.status, contentType: contentType};
+        const res: operations.CreateExpenseDatasetResponse = {statusCode: httpRes.status, contentType: contentType, rawResponse: httpRes};
         switch (true) {
           case httpRes?.status == 200:
             if (utils.matchContentType(contentType, `application/json`)) {
@@ -109,7 +109,7 @@ export class Expenses {
         const contentType: string = httpRes?.headers?.["content-type"] ?? "";
 
         if (httpRes?.status == null) throw new Error(`status code not found in response: ${httpRes}`);
-        const res: operations.UploadAttachmentResponse = {statusCode: httpRes.status, contentType: contentType};
+        const res: operations.UploadAttachmentResponse = {statusCode: httpRes.status, contentType: contentType, rawResponse: httpRes};
         switch (true) {
           case httpRes?.status == 200:
             if (utils.matchContentType(contentType, `application/json`)) {