+++ expenses/src/sdk/models/operations/getmappingoptions.ts
@@ -1,4 +1,5 @@
 import { SpeakeasyBase, SpeakeasyMetadata } from "../../../internal/utils";
+import { AxiosResponse } from "axios";
 import { Expose, Transform, Type } from "class-transformer";
 
 
@@ -137,6 +138,9 @@ export class GetMappingOptionsResponse extends SpeakeasyBase {
 
   @SpeakeasyMetadata()
   statusCode: number;
+
+  @SpeakeasyMetadata()
+  rawResponse?: AxiosResponse;
 
   @SpeakeasyMetadata()
   getMappingOptions200ApplicationJSONObject?: GetMappingOptions200ApplicationJSON;