+++ expenses/src/sdk/models/operations/savecompanyconfiguration.ts
@@ -1,4 +1,5 @@
 import { SpeakeasyBase, SpeakeasyMetadata } from "../../../internal/utils";
+import { AxiosResponse } from "axios";
 import { Expose, Type } from "class-transformer";
 
 
@@ -209,6 +210,9 @@ export class SaveCompanyConfigurationResponse extends SpeakeasyBase {
 
   @SpeakeasyMetadata()
   statusCode: number;
+
+  @SpeakeasyMetadata()
+  rawResponse?: AxiosResponse;
 
   @SpeakeasyMetadata()
   saveCompanyConfiguration200ApplicationJSONObject?: SaveCompanyConfiguration200ApplicationJSON;