+++ expenses/src/sdk/models/shared/security.ts
@@ -3,5 +3,5 @@ import { SpeakeasyBase, SpeakeasyMetadata } from "../../../internal/utils";
 
 export class Security extends SpeakeasyBase {
   @SpeakeasyMetadata({ data: "security, scheme=true;type=apiKey;subtype=header;name=Authorization" })
-  authorization: string;
+  authHeader: string;
 }