+++ expenses/src/sdk/sync.ts
@@ -62,7 +62,7 @@ export class Sync {
         const contentType: string = httpRes?.headers?.["content-type"] ?? "";
 
         if (httpRes?.status == null) throw new Error(`status code not found in response: ${httpRes}`);
-        const res: operations.IntiateSyncResponse = {statusCode: httpRes.status, contentType: contentType};
+        const res: operations.IntiateSyncResponse = {statusCode: httpRes.status, contentType: contentType, rawResponse: httpRes};
         switch (true) {
           case httpRes?.status == 202:
             if (utils.matchContentType(contentType, `application/json`)) {