+++ expenses/src/sdk/syncstatus.ts
@@ -49,7 +49,7 @@ export class SyncStatus {
         const contentType: string = httpRes?.headers?.["content-type"] ?? "";
 
         if (httpRes?.status == null) throw new Error(`status code not found in response: ${httpRes}`);
-        const res: operations.GetLastSuccessfulSyncResponse = {statusCode: httpRes.status, contentType: contentType};
+        const res: operations.GetLastSuccessfulSyncResponse = {statusCode: httpRes.status, contentType: contentType, rawResponse: httpRes};
         switch (true) {
           case httpRes?.status == 200:
             if (utils.matchContentType(contentType, `application/json`)) {
@@ -96,7 +96,7 @@ export class SyncStatus {
         const contentType: string = httpRes?.headers?.["content-type"] ?? "";
 
         if (httpRes?.status == null) throw new Error(`status code not found in response: ${httpRes}`);
-        const res: operations.GetLatestSyncResponse = {statusCode: httpRes.status, contentType: contentType};
+        const res: operations.GetLatestSyncResponse = {statusCode: httpRes.status, contentType: contentType, rawResponse: httpRes};
         switch (true) {
           case httpRes?.status == 200:
             if (utils.matchContentType(contentType, `application/json`)) {
@@ -143,7 +143,7 @@ export class SyncStatus {
         const contentType: string = httpRes?.headers?.["content-type"] ?? "";
 
         if (httpRes?.status == null) throw new Error(`status code not found in response: ${httpRes}`);
-        const res: operations.GetSyncByIdResponse = {statusCode: httpRes.status, contentType: contentType};
+        const res: operations.GetSyncByIdResponse = {statusCode: httpRes.status, contentType: contentType, rawResponse: httpRes};
         switch (true) {
           case httpRes?.status == 200:
             if (utils.matchContentType(contentType, `application/json`)) {
@@ -190,7 +190,7 @@ export class SyncStatus {
         const contentType: string = httpRes?.headers?.["content-type"] ?? "";
 
         if (httpRes?.status == null) throw new Error(`status code not found in response: ${httpRes}`);
-        const res: operations.ListSyncsResponse = {statusCode: httpRes.status, contentType: contentType};
+        const res: operations.ListSyncsResponse = {statusCode: httpRes.status, contentType: contentType, rawResponse: httpRes};
         switch (true) {
           case httpRes?.status == 200:
             if (utils.matchContentType(contentType, `application/json`)) {