+++ expenses/RELEASES.md
@@ -15,3 +15,11 @@ Based on:
 - Speakeasy CLI 1.7.1 https://github.com/speakeasy-api/speakeasy
 ### Releases
 - [NPM v0.1.1] https://www.npmjs.com/package/@codat/expenses/v/0.1.1 - expenses
+
+## 2023-03-04 00:21:41
+### Changes
+Based on:
+- OpenAPI Doc prealpha https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Expenses.yaml
+- Speakeasy CLI 1.8.2 https://github.com/speakeasy-api/speakeasy
+### Releases
+- [NPM v0.2.0] https://www.npmjs.com/package/@codat/expenses/v/0.2.0 - expenses